// src/types/UserType.ts
export interface UserType {
    name: string,
    email: string,
    bio?: string | undefined;
    profilePicture?: string | undefined;
    mobileNumber?: string | undefined;
    country?: string | undefined;
    city?: string | undefined;
    postalCode?: string | undefined;
    CV_URL?: string | undefined;
}